data_sources/requirements.txt
# Requires Python >= 3.8

# Google Analytics and Search Console
google-analytics-data>=0.18.0,<1.0.0
google-auth>=2.23.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.1.1,<1.0.0
google-api-python-client>=2.100.0,<3.0.0

# HTTP and API requests
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0

# Configuration and environment
python-dotenv>=1.0.0,<2.0.0

# Caching
diskcache>=5.6.0,<6.0.0

# Date/time handling
python-dateutil>=2.8.2,<3.0.0

# NLP and text analysis
textstat>=0.7.3,<1.0.0

# Machine learning for clustering (lazy-loaded by keyword_analyzer)
scikit-learn>=1.3.0,<2.0.0

# Web scraping and parsing
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<6.0.0

# Markdown processing (for WordPress publisher)
markdown>=3.5.0,<4.0.0
